Added a texture atlas class, and modified code all around

+++ Source/Graphics/IndexBuffer.cpp
@@ -9,7 +9,7 @@ namespace Engine {
 			glGenBuffers(1, &id);
 		}
 
-		IndexBuffer::IndexBuffer(const unsigned int* data, size_t size) : count(size) {
+		IndexBuffer::IndexBuffer(const GLushort* data, size_t size) : count(size) {
 			glGenBuffers(1, &id);
 			bind();
 			glBufferData(GL_ELEMENT_ARRAY_BUFFER, count * sizeof(unsigned int), data, GL_STATIC_DRAW);
